add useTerminalCommands and useTerminalResize composables
